[INTERPRETER] * BUG: Fix base object structure size when compiling with gcc 4.3 on a 64   bits system.
git-svn-id: svn://localhost/gambas/branches/2.0@1480 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gambas.h
@@ -98,7 +98,7 @@ typedef
 typedef
   struct {
     GB_CLASS klass;
-    int ref;
+    intptr_t ref;
     }
   GB_BASE;
